Support non blocking mode driver
non-blocking mode driver is a wrapper of underlying log drivers like
awslogs and an intermediate buffer. Non-blocking driver consumes container
logs continuously and saves them to the buffer. The underlying driver
consumes logs from the intermediate buffer and sends them to the final
destination.

Saving logs to buffer and consuming logs from buffer are different
go routines. Thus in that case, if we have, for example, networking
issues sending logs to cloudwatch, the container performance will not
be blocked as the log messages are saved in the intermediate buffer.

The intermediate buffer size defaults to 1 megabytes and customer can
set it to a different size. If the buffer is full, the oldest logs saved
in the memory will be dropped.

Add verbose mode to input arguments. When this boolean option is enabled,
more messages would be sent to system journal for debugging purpose.