Handle long log messages in non-blocking mode
Previously in non-blocking mode, the logger was still using scanner
with split function of SplitLines, which will close the reader side
of container pipe if the token is too long for the scanner to handle,
i.e. when a single log message is larger than the size of container pipe,
the reader will crash and stop consuming logs from container pipe. Thus
the container performance will be blocked. This commit resolves this issue.