Modify setuid/setgid syscall and only applies it once in main goroutine
Previously due to a bug of setuid/setgid syscall in golang, these two
syscalls only apply to current goroutine instead of the whole process.
With the fix checked in, we are removing the redundant calls of these
functions in sub-goroutines.

Note that golang has not included this fix in a specific version, and
in order to take it effect, please build shim logger with go built from
source.

Golang issue:
https://github.com/golang/go/issues/1435
Golang commit of fix:
https://github.com/golang/go/commit/d1b1145cace8b968307f9311ff611e4bb810710c