Only pass splunk tag to logging config if specified
The default splunk tag used when parsing `SplunkTagKey` is an empty
string. When passing the value of this flag, it is only passed to the
config if it is different from this default (i.e. a non-empty string).
However, an empty string as the default creates conflict it has meaning
within the Docker context. Docker will remove the default tag format
`{{.ID}}` if an empty string is specified.
Therefore, empty strings should be allowed as values for `SplunkTagKey`.

Now the logger will pass the value of `SplunkTagKey` to the
logging config only if the flag was set, rather than by determining
whether the value is not equal to the default (an empty string).