ci: add Splunk basic e2e tests (#73)
*Description of changes:*

The test is not perfect but can be used for basic validation because the
containerd task is ran successfully with the Splunk shim logger. It is
nice to validate the logs in Splunk after it. It is not straightforward
to do that so opened issue to track it.
https://github.com/aws/shim-loggers-for-containerd/issues/74

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Signed-off-by: Ziwen Ning <ningziwe@amazon.com>